+++ src/Abstractions/CrestApps.OrchardCore.AI.Abstractions/AIToolDefinitions.cs
@@ -1,34 +1,6 @@
 using Microsoft.Extensions.AI;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace CrestApps.OrchardCore.AI;
-
-public static class ServiceCollectionExtensions
-{
-    public static IServiceCollection AddAITool<TTool>(this IServiceCollection services, string name, Action<AIToolDefinitionEntry> configure = null)
-        where TTool : AITool
-    {
-        ArgumentException.ThrowIfNullOrEmpty(name);
-
-        services.AddScoped<TTool>();
-        services.Configure<AIToolDefinitions>(o =>
-        {
-            o.Add<TTool>(name, configure);
-        });
-
-        return services;
-    }
-
-    public static IServiceCollection AddAIToolSource<TSource>(this IServiceCollection services, string source)
-        where TSource : class, IAIToolSource
-    {
-        services.AddScoped<TSource>();
-        services.AddScoped<IAIToolSource>(sp => sp.GetRequiredService<TSource>());
-        services.AddKeyedScoped<IAIToolSource>(source, (sp, key) => sp.GetRequiredService<TSource>());
-
-        return services;
-    }
-}
 
 public class AIToolDefinitions
 {
@@ -51,18 +23,4 @@ public class AIToolDefinitions
 
         _tools[name] = definition;
     }
-}
-
-public class AIToolDefinitionEntry
-{
-    public AIToolDefinitionEntry(Type type)
-    {
-        ToolType = type;
-    }
-
-    public Type ToolType { get; }
-
-    public string Title { get; set; }
-
-    public string Description { get; set; }
 }
