+++ src/Abstractions/CrestApps.OrchardCore.AI.Abstractions/AIToolSourceType.cs
@@ -1,0 +1,7 @@
+﻿namespace CrestApps.OrchardCore.AI;
+
+public enum AIToolSourceType
+{
+    Tool,
+    Function,
+}
