+++ src/Abstractions/CrestApps.OrchardCore.AI.Abstractions/Models/AIToolInstance.cs
@@ -34,4 +34,3 @@ public class AIToolInstance : Entity
         return DisplayText;
     }
 }
-
