+++ src/Abstractions/CrestApps.OrchardCore.AI.Abstractions/Models/ValidatedModelDeploymentContext.cs
@@ -2,9 +2,9 @@ namespace CrestApps.OrchardCore.AI.Models;
 
 public sealed class ValidatedModelDeploymentContext : AIDeploymentContextBase
 {
-    public readonly AIValidateResult Result;
+    public readonly ValidationResultDetails Result;
 
-    public ValidatedModelDeploymentContext(AIDeployment deployment, AIValidateResult result)
+    public ValidatedModelDeploymentContext(AIDeployment deployment, ValidationResultDetails result)
         : base(deployment)
     {
         Result = result ?? new();
