+++ src/Abstractions/CrestApps.OrchardCore.AI.Abstractions/Models/ValidatingAIDeploymentContext.cs
@@ -2,7 +2,7 @@ namespace CrestApps.OrchardCore.AI.Models;
 
 public sealed class ValidatingAIDeploymentContext : AIDeploymentContextBase
 {
-    public AIValidateResult Result { get; } = new();
+    public ValidationResultDetails Result { get; }
 
     public ValidatingAIDeploymentContext(AIDeployment deployment)
         : base(deployment)
