+++ src/Abstractions/CrestApps.OrchardCore.AI.Abstractions/Models/ValidatingAIProfileContext.cs
@@ -2,7 +2,7 @@ namespace CrestApps.OrchardCore.AI.Models;
 
 public sealed class ValidatingAIProfileContext : AIProfileContextBase
 {
-    public AIValidateResult Result { get; } = new();
+    public ValidationResultDetails Result { get; } = new();
 
     public ValidatingAIProfileContext(AIProfile profile)
         : base(profile)
