+++ src/Abstractions/CrestApps.OrchardCore.AI.Abstractions/Models/ValidatingAIToolInstanceContext.cs
@@ -2,7 +2,7 @@
 
 public sealed class ValidatingAIToolInstanceContext : AIToolInstanceContextBase
 {
-    public AIValidateResult Result { get; } = new();
+    public ValidationResultDetails Result { get; } = new();
 
     public ValidatingAIToolInstanceContext(AIToolInstance instance)
         : base(instance)
