+++ src/Abstractions/CrestApps.OrchardCore.AI.Abstractions/ServiceCollectionExtensions.cs
@@ -1,0 +1,31 @@
+﻿using Microsoft.Extensions.AI;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace CrestApps.OrchardCore.AI;
+
+public static class ServiceCollectionExtensions
+{
+    public static IServiceCollection AddAITool<TTool>(this IServiceCollection services, string name, Action<AIToolDefinitionEntry> configure = null)
+        where TTool : AITool
+    {
+        ArgumentException.ThrowIfNullOrEmpty(name);
+
+        services.AddScoped<TTool>();
+        services.Configure<AIToolDefinitions>(o =>
+        {
+            o.Add<TTool>(name, configure);
+        });
+
+        return services;
+    }
+
+    public static IServiceCollection AddAIToolSource<TSource>(this IServiceCollection services, string source)
+        where TSource : class, IAIToolSource
+    {
+        services.AddScoped<TSource>();
+        services.AddScoped<IAIToolSource>(sp => sp.GetRequiredService<TSource>());
+        services.AddKeyedScoped<IAIToolSource>(source, (sp, key) => sp.GetRequiredService<TSource>());
+
+        return services;
+    }
+}
