+++ src/Startup/CrestApps.OrchardCore.Cms.Web/CrestApps.OrchardCore.Cms.Web.csproj
@@ -9,7 +9,6 @@
 
   <ItemGroup>
     <Folder Include="wwwroot\" />
-    <Folder Include="Localization\" />
   </ItemGroup>
 
   <!-- Watcher include and excludes -->
