--- src/Startup/CrestApps.OrchardCore.Cms.Web/Localization/.placeholder
@@ -1,1 +1,0 @@
-ÿþ
